Implement address creation functionality in API
This commit includes the changes necessary to introduce the ability to add addresses to a contact in the API. Necessary changes have been made to the database schema, and new routes, validation, service methods and controllers have been added to handle address creation requests. In addition, tests are implemented to validate the new functionality provided. The postal_cde column in the Address model has been dropped and replaced with postal_code.