Add getAddress functionality and related tests
This commit provides support for retrieving a contact's address by adding a new "getAddress" function in the API. It comprises an update to routes and models, the creation of new validation rules and service methods, plus an enhancement of the AddressController. Also, comprehensive "getAddress" tests have been added to ensure the function works as expected.