Add list address functionality and associated tests
This commit introduces a 'list' function allowing to retrieve a list of addresses for a specific contact. It includes updates to the 'api.ts' route, adds the respective service in 'address-service.ts', and implements a new controller in 'address-controller.ts'. Corresponding unit tests have also been added for ensuring the proper functionality of the new endpoint.